requirement.txt
flask
pandas
joblib
fuzzywuzzy
nltk.tokenize
sklearn.feature_extraction.text
joblib